+++ src/components/checkout/index.ts
@@ -2,3 +2,6 @@ export { CheckoutStepper } from "./CheckoutStepper";
 export { CheckoutAddressStep } from "./CheckoutAddressStep";
 export { CheckoutOrderReview } from "./CheckoutOrderReview";
 export { OrderSummaryPanel } from "./OrderSummaryPanel";
+export { OrderSuccessHero } from "./OrderSuccessHero";
+export { OrderSuccessCard } from "./OrderSuccessCard";
+export { OrderSuccessActions } from "./OrderSuccessActions";